// src/types/wikipediaTypes.ts
export interface DataFromWiki {
  type: string
  title: string
  displaytitle: string
  namespace: Namespace
  wikibase_item: string
  titles: Titles
  pageid: number
  thumbnail: Thumbnail
  originalimage: Originalimage
  lang: string
  dir: string
  revision: string
  tid: string
  timestamp: string
  description: string
  description_source: string
  coordinates: Coordinates
  content_urls: ContentUrls
  extract: string
  extract_html: string
}

export interface Namespace {
  id: number
  text: string
}

export interface Titles {
  canonical: string
  normalized: string
  display: string
}

export interface Thumbnail {
  source: string
  width: number
  height: number
}

export interface Originalimage {
  source: string
  width: number
  height: number
}

export interface Coordinates {
  lat: number
  lon: number
}

export interface ContentUrls {
  desktop: Desktop
  mobile: Mobile
}

export interface Desktop {
  page: string
  revisions: string
  edit: string
  talk: string
}

export interface Mobile {
  page: string
  revisions: string
  edit: string
  talk: string
}